simplify guardian review flow
