+++ codex-rs/core/src/guardian_tests.rs
@@ -1,0 +1,498 @@
+use super::*;
+use crate::config::ManagedFeatures;
+use crate::config::NetworkProxySpec;
+use crate::config::test_config;
+use crate::config_loader::FeatureRequirementsToml;
+use crate::config_loader::NetworkConstraints;
+use crate::config_loader::RequirementSource;
+use crate::config_loader::Sourced;
+use crate::test_support;
+use codex_network_proxy::NetworkProxyConfig;
+use codex_protocol::models::ContentItem;
+use core_test_support::context_snapshot;
+use core_test_support::context_snapshot::ContextSnapshotOptions;
+use core_test_support::responses::ev_assistant_message;
+use core_test_support::responses::ev_completed;
+use core_test_support::responses::ev_response_created;
+use core_test_support::responses::mount_sse_once;
+use core_test_support::responses::sse;
+use core_test_support::responses::start_mock_server;
+use core_test_support::skip_if_no_network;
+use insta::assert_snapshot;
+use pretty_assertions::assert_eq;
+use std::collections::BTreeMap;
+use std::path::PathBuf;
+use std::sync::Arc;
+use std::time::Duration;
+use tokio_util::sync::CancellationToken;
+
+#[test]
+fn build_guardian_transcript_keeps_original_numbering() {
+    let entries = [
+        GuardianTranscriptEntry {
+            number: 1,
+            role: "user".to_string(),
+            is_user: true,
+            is_tool: false,
+            text: "first".to_string(),
+        },
+        GuardianTranscriptEntry {
+            number: 2,
+            role: "assistant".to_string(),
+            is_user: false,
+            is_tool: false,
+            text: "second".to_string(),
+        },
+        GuardianTranscriptEntry {
+            number: 3,
+            role: "assistant".to_string(),
+            is_user: false,
+            is_tool: false,
+            text: "third".to_string(),
+        },
+    ];
+
+    let (transcript, omission) = build_guardian_transcript(&entries[..2]);
+
+    assert!(transcript.contains("[1] user: first"));
+    assert!(transcript.contains("[2] assistant: second"));
+    assert!(omission.is_none());
+}
+
+#[test]
+fn collect_guardian_transcript_entries_skips_contextual_user_messages() {
+    let items = vec![
+        ResponseItem::Message {
+            id: None,
+            role: "user".to_string(),
+            content: vec![ContentItem::InputText {
+                text: "<environment_context>\n<cwd>/tmp</cwd>\n</environment_context>".to_string(),
+            }],
+            end_turn: None,
+            phase: None,
+        },
+        ResponseItem::Message {
+            id: None,
+            role: "assistant".to_string(),
+            content: vec![ContentItem::OutputText {
+                text: "hello".to_string(),
+            }],
+            end_turn: None,
+            phase: None,
+        },
+    ];
+
+    let entries = collect_guardian_transcript_entries(&items);
+
+    assert_eq!(entries.len(), 1);
+    assert_eq!(entries[0].number, 1);
+    assert_eq!(entries[0].role, "assistant");
+}
+
+#[test]
+fn collect_guardian_transcript_entries_includes_recent_tool_calls_and_output() {
+    let items = vec![
+        ResponseItem::Message {
+            id: None,
+            role: "user".to_string(),
+            content: vec![ContentItem::InputText {
+                text: "check the repo".to_string(),
+            }],
+            end_turn: None,
+            phase: None,
+        },
+        ResponseItem::FunctionCall {
+            id: None,
+            name: "read_file".to_string(),
+            arguments: "{\"path\":\"README.md\"}".to_string(),
+            call_id: "call-1".to_string(),
+        },
+        ResponseItem::FunctionCallOutput {
+            call_id: "call-1".to_string(),
+            output: codex_protocol::models::FunctionCallOutputPayload::from_text(
+                "repo is public".to_string(),
+            ),
+        },
+        ResponseItem::Message {
+            id: None,
+            role: "assistant".to_string(),
+            content: vec![ContentItem::OutputText {
+                text: "I need to push a fix".to_string(),
+            }],
+            end_turn: None,
+            phase: None,
+        },
+    ];
+
+    let entries = collect_guardian_transcript_entries(&items);
+
+    assert_eq!(entries.len(), 4);
+    assert_eq!(entries[1].role, "tool read_file call");
+    assert!(entries[1].is_tool);
+    assert_eq!(entries[1].text, "{\"path\":\"README.md\"}");
+    assert_eq!(entries[2].role, "tool read_file result");
+    assert_eq!(entries[2].text, "repo is public");
+}
+
+#[test]
+fn guardian_truncate_text_keeps_prefix_suffix_and_xml_marker() {
+    let content = "prefix ".repeat(200) + &" suffix".repeat(200);
+
+    let truncated = guardian_truncate_text(&content, 20);
+
+    assert!(truncated.starts_with("prefix"));
+    assert!(truncated.contains("<guardian_truncated omitted_approx_tokens=\""));
+    assert!(truncated.ends_with("suffix"));
+}
+
+#[test]
+fn format_guardian_action_pretty_truncates_large_string_fields() {
+    let action = serde_json::json!({
+        "tool": "apply_patch",
+        "cwd": PathBuf::from("/tmp"),
+        "files": Vec::<String>::new(),
+        "change_count": 1usize,
+        "patch": "line\n".repeat(2_000),
+    });
+
+    let rendered = format_guardian_action_pretty(&action);
+
+    assert!(rendered.contains("\"tool\": \"apply_patch\""));
+    assert!(rendered.contains("<guardian_truncated omitted_approx_tokens=\""));
+}
+
+#[test]
+fn build_guardian_transcript_reserves_separate_budget_for_tool_evidence() {
+    let repeated = "signal ".repeat(8_000);
+    let entries = vec![
+        GuardianTranscriptEntry {
+            number: 1,
+            role: "user".to_string(),
+            is_user: true,
+            is_tool: false,
+            text: "please figure out if the repo is public".to_string(),
+        },
+        GuardianTranscriptEntry {
+            number: 2,
+            role: "tool gh".to_string(),
+            is_user: false,
+            is_tool: true,
+            text: repeated.clone(),
+        },
+        GuardianTranscriptEntry {
+            number: 3,
+            role: "tool read_file".to_string(),
+            is_user: false,
+            is_tool: true,
+            text: repeated.clone(),
+        },
+        GuardianTranscriptEntry {
+            number: 4,
+            role: "tool web".to_string(),
+            is_user: false,
+            is_tool: true,
+            text: repeated.clone(),
+        },
+        GuardianTranscriptEntry {
+            number: 5,
+            role: "tool gh".to_string(),
+            is_user: false,
+            is_tool: true,
+            text: repeated,
+        },
+        GuardianTranscriptEntry {
+            number: 6,
+            role: "assistant".to_string(),
+            is_user: false,
+            is_tool: false,
+            text: "The public repo check is the main reason I want to escalate.".to_string(),
+        },
+    ];
+
+    let (transcript, omission) = build_guardian_transcript(&entries);
+
+    assert!(transcript.contains("[1] user: please figure out if the repo is public"));
+    assert!(
+        transcript.contains(
+            "[6] assistant: The public repo check is the main reason I want to escalate."
+        )
+    );
+    assert!(!transcript.contains("[2] tool gh:"));
+    assert!(omission.is_some());
+}
+
+#[test]
+fn parse_guardian_assessment_extracts_embedded_json() {
+    let parsed = parse_guardian_assessment(Some(
+        "preface {\"risk_level\":\"medium\",\"risk_score\":42,\"rationale\":\"ok\",\"evidence\":[]}",
+    ))
+    .expect("guardian assessment");
+
+    assert_eq!(parsed.risk_score, 42);
+    assert_eq!(parsed.risk_level, GuardianRiskLevel::Medium);
+}
+
+#[tokio::test(flavor = "multi_thread", worker_threads = 2)]
+async fn guardian_review_request_layout_matches_model_visible_request_snapshot()
+-> anyhow::Result<()> {
+    skip_if_no_network!(Ok(()));
+
+    let server = start_mock_server().await;
+    let guardian_assessment = serde_json::json!({
+        "risk_level": "medium",
+        "risk_score": 35,
+        "rationale": "The user explicitly requested pushing the reviewed branch to the known remote.",
+        "evidence": [{
+            "message": "The user asked to check repo visibility and then push the docs fix.",
+            "why": "This authorizes the specific network action under review.",
+        }],
+    })
+    .to_string();
+    let request_log = mount_sse_once(
+        &server,
+        sse(vec![
+            ev_response_created("resp-guardian"),
+            ev_assistant_message("msg-guardian", &guardian_assessment),
+            ev_completed("resp-guardian"),
+        ]),
+    )
+    .await;
+
+    let (mut session, mut turn) = crate::codex::make_session_and_context().await;
+    let mut config = (*turn.config).clone();
+    config.model_provider.base_url = Some(format!("{}/v1", server.uri()));
+    let config = Arc::new(config);
+    let models_manager = Arc::new(test_support::models_manager_with_provider(
+        config.codex_home.clone(),
+        Arc::clone(&session.services.auth_manager),
+        config.model_provider.clone(),
+    ));
+    session.services.models_manager = models_manager;
+    turn.config = Arc::clone(&config);
+    turn.provider = config.model_provider.clone();
+    let session = Arc::new(session);
+    let turn = Arc::new(turn);
+
+    session
+        .record_into_history(
+            &[
+                ResponseItem::Message {
+                    id: None,
+                    role: "user".to_string(),
+                    content: vec![ContentItem::InputText {
+                        text: "Please check the repo visibility and push the docs fix if needed."
+                            .to_string(),
+                    }],
+                    end_turn: None,
+                    phase: None,
+                },
+                ResponseItem::FunctionCall {
+                    id: None,
+                    name: "gh_repo_view".to_string(),
+                    arguments: "{\"repo\":\"openai/codex\"}".to_string(),
+                    call_id: "call-1".to_string(),
+                },
+                ResponseItem::FunctionCallOutput {
+                    call_id: "call-1".to_string(),
+                    output: codex_protocol::models::FunctionCallOutputPayload::from_text(
+                        "repo visibility: public".to_string(),
+                    ),
+                },
+                ResponseItem::Message {
+                    id: None,
+                    role: "assistant".to_string(),
+                    content: vec![ContentItem::OutputText {
+                        text: "The repo is public; I now need approval to push the docs fix."
+                            .to_string(),
+                    }],
+                    end_turn: None,
+                    phase: None,
+                },
+            ],
+            turn.as_ref(),
+        )
+        .await;
+
+    let prompt = build_guardian_prompt_items(
+        session.as_ref(),
+        Some("Sandbox denied outbound git push to github.com.".to_string()),
+        GuardianReviewRequest {
+            action: serde_json::json!({
+                "tool": "shell",
+                "command": [
+                    "git",
+                    "push",
+                    "origin",
+                    "guardian-approval-mvp"
+                ],
+                "cwd": "/repo/codex-rs/core",
+                "sandbox_permissions": crate::sandboxing::SandboxPermissions::UseDefault,
+                "justification": "Need to push the reviewed docs fix to the repo remote.",
+            }),
+        },
+    )
+    .await;
+
+    let assessment = run_guardian_subagent(
+        Arc::clone(&session),
+        Arc::clone(&turn),
+        prompt,
+        guardian_output_schema(),
+        CancellationToken::new(),
+    )
+    .await?;
+    assert_eq!(assessment.risk_score, 35);
+
+    let request = request_log.single_request();
+    assert_snapshot!(
+        "guardian_review_request_layout",
+        context_snapshot::format_labeled_requests_snapshot(
+            "Guardian review request layout",
+            &[("Guardian Review Request", &request)],
+            &ContextSnapshotOptions::default(),
+        )
+    );
+
+    Ok(())
+}
+
+#[tokio::test]
+async fn guardian_timeout_cancels_subagent_token() {
+    let cancel_token = CancellationToken::new();
+    let waiter = tokio::spawn({
+        let cancel_token = cancel_token.clone();
+        async move {
+            cancel_token.cancelled().await;
+        }
+    });
+
+    let result = run_guardian_subagent_with_timeout(
+        std::future::pending::<anyhow::Result<GuardianAssessment>>(),
+        cancel_token,
+        Duration::from_millis(10),
+    )
+    .await;
+
+    assert!(matches!(result, Err(GuardianReviewFailure::TimedOut)));
+    tokio::time::timeout(Duration::from_secs(1), waiter)
+        .await
+        .expect("timeout helper should cancel the guardian token")
+        .expect("waiter task should finish cleanly");
+}
+
+#[test]
+fn guardian_subagent_config_preserves_parent_network_proxy() {
+    let mut parent_config = test_config();
+    let network = NetworkProxySpec::from_config_and_constraints(
+        NetworkProxyConfig::default(),
+        Some(NetworkConstraints {
+            enabled: Some(true),
+            allowed_domains: Some(vec!["github.com".to_string()]),
+            ..Default::default()
+        }),
+        parent_config.permissions.sandbox_policy.get(),
+    )
+    .expect("network proxy spec");
+    parent_config.permissions.network = Some(network.clone());
+
+    let guardian_config = build_guardian_subagent_config(
+        &parent_config,
+        None,
+        "parent-active-model",
+        Some(codex_protocol::openai_models::ReasoningEffort::Low),
+    )
+    .expect("guardian config");
+
+    assert_eq!(guardian_config.permissions.network, Some(network));
+    assert_eq!(
+        guardian_config.model,
+        Some("parent-active-model".to_string())
+    );
+    assert_eq!(
+        guardian_config.model_reasoning_effort,
+        Some(codex_protocol::openai_models::ReasoningEffort::Low)
+    );
+    assert_eq!(
+        guardian_config.permissions.approval_policy,
+        Constrained::allow_only(AskForApproval::Never)
+    );
+    assert_eq!(
+        guardian_config.permissions.sandbox_policy,
+        Constrained::allow_only(SandboxPolicy::new_read_only_policy())
+    );
+}
+
+#[test]
+fn guardian_subagent_config_uses_live_network_proxy_state() {
+    let mut parent_config = test_config();
+    let mut parent_network = NetworkProxyConfig::default();
+    parent_network.network.enabled = true;
+    parent_network.network.allowed_domains = vec!["parent.example".to_string()];
+    parent_config.permissions.network = Some(
+        NetworkProxySpec::from_config_and_constraints(
+            parent_network,
+            None,
+            parent_config.permissions.sandbox_policy.get(),
+        )
+        .expect("parent network proxy spec"),
+    );
+
+    let mut live_network = NetworkProxyConfig::default();
+    live_network.network.enabled = true;
+    live_network.network.allowed_domains = vec!["github.com".to_string()];
+
+    let guardian_config = build_guardian_subagent_config(
+        &parent_config,
+        Some(live_network.clone()),
+        "active-model",
+        None,
+    )
+    .expect("guardian config");
+
+    assert_eq!(
+        guardian_config.permissions.network,
+        Some(
+            NetworkProxySpec::from_config_and_constraints(
+                live_network,
+                None,
+                &SandboxPolicy::new_read_only_policy(),
+            )
+            .expect("live network proxy spec")
+        )
+    );
+}
+
+#[test]
+fn guardian_subagent_config_rejects_pinned_collab_feature() {
+    let mut parent_config = test_config();
+    parent_config.features = ManagedFeatures::from_configured(
+        parent_config.features.get().clone(),
+        Some(Sourced {
+            value: FeatureRequirementsToml {
+                entries: BTreeMap::from([("multi_agent".to_string(), true)]),
+            },
+            source: RequirementSource::Unknown,
+        }),
+    )
+    .expect("managed features");
+
+    let err = build_guardian_subagent_config(&parent_config, None, "active-model", None)
+        .expect_err("guardian config should fail when collab is pinned on");
+
+    assert!(
+        err.to_string()
+            .contains("guardian subagent requires `features.multi_agent` to be disabled")
+    );
+}
+
+#[test]
+fn guardian_subagent_config_uses_parent_active_model_instead_of_hardcoded_slug() {
+    let mut parent_config = test_config();
+    parent_config.model = Some("configured-model".to_string());
+
+    let guardian_config =
+        build_guardian_subagent_config(&parent_config, None, "active-model", None)
+            .expect("guardian config");
+
+    assert_eq!(guardian_config.model, Some("active-model".to_string()));
+}
