+++ codex-rs/core/src/tools/runtimes/shell.rs
@@ -170,10 +170,7 @@ impl Approvable<ShellRequest> for ShellRuntime {
                         action.remove("justification");
                     }
                 }
-                let request = GuardianReviewRequest {
-                    tool_name: "shell",
-                    action,
-                };
+                let request = GuardianReviewRequest { action };
                 return review_approval_request(session, turn, request).await;
             }
             with_cached_approval(&session.services, "shell", keys, move || async move {
